+++ src/Contract/RouteFindByTargetsAndTypeActionInterface.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\Base\Contract;
+
+interface RouteFindByTargetsAndTypeActionInterface
+{
+    public function find(RouteFindByTargetsAndTypeCriteria $criteria): ?RouteFindByTargetsAndTypeResult;
+}